Refatoração completa da API.

+++ src/main/java/com/bytedev/storage/service/ProductService.java
@@ -1,8 +1,6 @@
 package com.bytedev.storage.service;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,14 +8,11 @@ import org.springframework.stereotype.Service;
 
 import com.bytedev.storage.domain.Category;
 import com.bytedev.storage.domain.Product;
-import com.bytedev.storage.domain.ProductStorage;
-import com.bytedev.storage.domain.Storage;
 import com.bytedev.storage.dto.ProductDTO;
-import com.bytedev.storage.dto.ProductStorageDTO;
+import com.bytedev.storage.exception.CustomException;
+import com.bytedev.storage.exception.CustomNotFoundException;
 import com.bytedev.storage.repository.CategoryRepository;
 import com.bytedev.storage.repository.ProductRepository;
-import com.bytedev.storage.repository.ProductStorageRepository;
-import com.bytedev.storage.repository.StorageRepository;
 
 import lombok.AllArgsConstructor;
 
@@ -30,51 +25,56 @@ public class ProductService {
     @Autowired
     private final CategoryRepository categoryRepository;
 
-    public List<ProductDTO> listProductWithCategoryStorage() {
-        return productRepository.findAll().stream()
-                .map(ProductDTO::new)
-                .collect(Collectors.toList());
+    public List<ProductDTO> findAll() {
+        try {
+            return productRepository.findAll().stream()
+                    .map(ProductDTO::new)
+                    .collect(Collectors.toList());
+        } catch (Exception e) {
+            throw new CustomException("Error fetching products: " + e.getMessage());
+        }
     }
 
-    public ProductDTO getProductById(Long id) {
+    public ProductDTO findById(Long id) {
         return productRepository.findById(id)
                 .map(ProductDTO::new)
-                .orElse(null);
+                .orElseThrow(() -> new RuntimeException(
+                        "Product not found: " + id));
     }
 
-    public ProductDTO saveProduct(ProductDTO productDTO) {
-        Product product = new Product();
-        product.setName(productDTO.getName());
-        product.setPrice(productDTO.getPrice());
-
-        if (productDTO.getCategoryId() != null) {
-            Category category = categoryRepository.findById(productDTO.getCategoryId())
-                    .orElseThrow(() -> new RuntimeException("Category not found: " + productDTO.getCategoryId()));
+    public ProductDTO create(ProductDTO dto) {
+        Product product = dto.toEntity();
+        if (dto.getCategoryId() != null) {
+            Category category = categoryRepository.findById(dto.getCategoryId())
+                    .orElseThrow(() -> new RuntimeException("Category not found: " + dto.getCategoryId()));
             product.setCategory(category);
         }
-
         Product savedProduct = productRepository.save(product);
         return new ProductDTO(savedProduct);
     }
 
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
-    }
-
-    public ProductDTO updateProduct(Long id, ProductDTO productDTO) {
+    public ProductDTO update(Long id, ProductDTO dto) {
         return productRepository.findById(id)
                 .map(product -> {
-                    product.setName(productDTO.getName());
-                    product.setPrice(productDTO.getPrice());
-                    
-                    if (productDTO.getCategoryId() != null) {
-                        Category category = categoryRepository.findById(productDTO.getCategoryId())
-                                .orElseThrow(() -> new RuntimeException("Category not found: " + productDTO.getCategoryId()));
+                    dto.toEntity();
+                    if (dto.getCategoryId() != null) {
+                        Category category = categoryRepository.findById(dto.getCategoryId())
+                                .orElseThrow(() -> new RuntimeException("Category not found: " + dto.getCategoryId()));
                         product.setCategory(category);
                     }
-                    
                     return new ProductDTO(productRepository.save(product));
                 })
                 .orElseThrow(() -> new RuntimeException("Product not found: " + id));
+    }
+
+    public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new CustomNotFoundException("Product not found with id: " + id);
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new CustomException("Error deleting product: " + e.getMessage());
+        }
     }
 }
